TASK 1:- CALCULATOR
LEVEL 2

TASK 1:- CALCULATOR

Created a responsive Calculator with basic mathematical functionalities using CSS, HTML, and JavaScript. It has an interactive interface to perform basic functions such as addition, subtraction, division, and multiplication and there's also a display screen to display the user’s input and give the results. The grid system in CSS is for the alignments of buttons on the calculator.
#oasisinfobyte
#aicteinternship
#webdevelopment