TASK 3:- A BASIC TO-DO WEBAPP
LEVEL 2
TASK 3:- A BASIC TO-DO WEBAPP

Developed a basic to-do app that enables a user to add daily tasks.

This to-do web app allows users to add new items to a list, and items added to a list must be visible. Also, a user can mark a task “complete” upon its successful completion, and cancel after being completed.